refactor: decompose 6 oversized components into smaller modules
Split large single-file components into focused sub-components:

- AccountDetails (2622→1107 lines): AccountHeader, TransactionForm,
  PendingOccurrences, TransactionRow, AutocompleteInput
- Dashboard (1821→1504 lines): TimeRangeSelector, AccountFilterPopover,
  SummaryCards, NetWorthChart, AssetAllocationChart, ExpensesByCategoryChart,
  IncomeVsExpensesChart
- SettingsView (1246→259 lines): GeneralSection, CustomizationSection,
  FormatsSection, AboutSection, LlmSettingsSection
- ScheduledList (1207→379 lines): ScheduledForm, ScheduledTable
- FireCalculator (1222→665 lines): FireInputsPanel, FireResultsPanel
- ImportModal (1082→849 lines): FileDropZone, ColumnMappingStep

All state and logic remains in parent components; children are presentational.
All 358 tests pass, lint clean (0 errors).
