+++ app/src/features/fire/FireCalculator.tsx
@@ -1,22 +1,5 @@
 import { useState, useEffect, useMemo, useRef, useCallback } from "react";
 import { rust } from "../../api/tauri-client";
-import { Line } from "react-chartjs-2";
-import {
-  TrendingUp,
-  Banknote,
-  Percent,
-  Calendar,
-  RotateCw,
-  ChevronDown,
-  ChevronRight,
-  User,
-  Target,
-  Clock,
-  Activity,
-  Settings2,
-} from "lucide-react";
-import { useFormatNumber } from "../../utils/format";
-import MaskedNumber from "../../components/ui/MaskedNumber";
 import {
   runMonteCarloSimulation,
   calculateDeterministicProjection,
@@ -29,7 +12,6 @@ import {
   computePortfolioTotals,
   computeNetWorthMarketValues,
 } from "../../utils/investments";
-import NumberInput from "../../components/ui/NumberInput";
 import {
   ErrorState,
   SkeletonCard,
@@ -47,6 +29,15 @@ import {
   Legend,
   Filler,
 } from "chart.js";
+import type {
+  Account,
+  InvestmentTransaction,
+  InvestmentQuote,
+  ProjectionResult,
+  MonteCarloResult,
+} from "./fire-types";
+import FireInputsPanel from "./FireInputsPanel";
+import FireResultsPanel from "./FireResultsPanel";
 
 ChartJS.register(
   CategoryScale,
@@ -58,50 +49,6 @@ ChartJS.register(
   Legend,
   Filler,
 );
-
-interface Account {
-  id: number;
-  kind: string;
-  balance: number;
-  [key: string]: unknown;
-}
-
-interface InvestmentTransaction {
-  date: string;
-  ticker?: string;
-  shares?: number;
-  price_per_share?: number;
-  fee?: number;
-  account_id: number;
-  amount: number;
-  category?: string;
-  [key: string]: unknown;
-}
-
-interface InvestmentQuote {
-  symbol: string;
-  regularMarketPrice: number;
-  [key: string]: unknown;
-}
-
-interface ProjectionResult {
-  fireNumber: number;
-  yearsToFire: number | null;
-  projectionData: number[];
-  neverReached: boolean;
-}
-
-interface MonteCarloResult {
-  successRate: number;
-  simulationCount: number;
-  percentiles: {
-    p10: number[];
-    p25: number[];
-    p50: number[];
-    p75: number[];
-    p90: number[];
-  };
-}
 
 export default function FireCalculator() {
   const { t } = useTranslation();
@@ -615,8 +562,6 @@ export default function FireCalculator() {
   // Calculate retirement age when FIRE is reached
   const fireAge = yearsToFire ? currentAge + yearsToFire : null;
 
-  const formatNumber = useFormatNumber();
-
   if (loading) {
     return (
       <div className="page-container fire-calculator-container animate-pulse">
@@ -674,547 +619,46 @@ export default function FireCalculator() {
 
       <div className="grid grid-cols-1 lg:grid-cols-4 gap-8">
         {/* Inputs */}
-        <div className="bg-white dark:bg-slate-800 p-6 rounded-2xl shadow-md border border-slate-200 dark:border-slate-700 space-y-6 h-fit hover:shadow-lg transition-shadow duration-300">
-          <div className="flex items-center justify-between mb-4">
-            <h2 className="text-xl font-bold text-slate-900 dark:text-slate-100">
-              {t("fire.parameters")}
-            </h2>
-            <button
-              type="button"
-              onClick={resetToHistoric}
-              title={t("fire.reset_tooltip")}
-              className="inline-flex items-center gap-2 px-3 py-2 text-sm font-medium text-slate-700 dark:text-slate-300 bg-slate-100 dark:bg-slate-700/30 border border-slate-200 dark:border-slate-600 rounded-lg hover:bg-slate-200 dark:hover:bg-slate-700 transition"
-            >
-              <RotateCw className="w-4 h-4" />
-              {t("fire.reset")}
-            </button>
-          </div>
-
-          <div className="space-y-5">
-            {/* Financial Parameters */}
-            <div>
-              <label className="form-label !mb-2 flex items-center gap-2">
-                <Banknote className="w-4 h-4 text-brand-500 dark:text-brand-400" />
-                {t("dashboard.current_net_worth")}
-              </label>
-              <div className="relative">
-                <NumberInput
-                  value={currentNetWorth}
-                  onChange={(num) => {
-                    setCurrentNetWorth(Number.isNaN(num) ? 0 : Math.round(num));
-                    userModified.current.currentNetWorth = true;
-                  }}
-                  className="form-input-lg font-semibold"
-                  placeholder="0"
-                  maximumFractionDigits={0}
-                  minimumFractionDigits={0}
-                  useGrouping={false}
-                />
-              </div>
-            </div>
-
-            <div>
-              <label className="form-label !mb-2 flex items-center gap-2">
-                <Banknote className="w-4 h-4 text-brand-500 dark:text-brand-400" />
-                {t("fire.annual_expenses")}
-              </label>
-              <div className="relative">
-                <NumberInput
-                  value={annualExpenses}
-                  onChange={(num) => {
-                    setAnnualExpenses(Number.isNaN(num) ? 0 : Math.round(num));
-                    userModified.current.annualExpenses = true;
-                  }}
-                  className="form-input-lg font-semibold"
-                  placeholder="0"
-                  maximumFractionDigits={0}
-                  minimumFractionDigits={0}
-                  useGrouping={false}
-                />
-              </div>
-            </div>
-
-            <div>
-              <label className="form-label !mb-2 flex items-center gap-2">
-                <Banknote className="w-4 h-4 text-brand-500 dark:text-brand-400" />
-                {t("fire.annual_savings")}
-              </label>
-              <div className="relative">
-                <NumberInput
-                  value={annualSavings}
-                  onChange={(num) => {
-                    setAnnualSavings(Number.isNaN(num) ? 0 : Math.round(num));
-                    userModified.current.annualSavings = true;
-                  }}
-                  className="form-input-lg font-semibold"
-                  placeholder="0"
-                  maximumFractionDigits={0}
-                  minimumFractionDigits={0}
-                  useGrouping={false}
-                />
-              </div>
-            </div>
-
-            <div>
-              <label className="form-label !mb-2 flex items-center gap-2">
-                <Percent className="w-4 h-4 text-brand-500 dark:text-brand-400" />
-                {t("fire.expected_return")}
-              </label>
-              <div className="relative">
-                <NumberInput
-                  value={expectedReturn}
-                  onChange={(num) => {
-                    setExpectedReturn(Number.isNaN(num) ? 0 : num);
-                    userModified.current.expectedReturn = true;
-                  }}
-                  className="form-input-lg font-semibold pr-8"
-                  placeholder="0"
-                  maximumFractionDigits={2}
-                  minimumFractionDigits={0}
-                  useGrouping={false}
-                />
-                <span className="absolute right-4 top-1/2 -translate-y-1/2 text-slate-400 font-medium">
-                  %
-                </span>
-              </div>
-            </div>
-
-            <div>
-              <label className="form-label !mb-2 flex items-center gap-2">
-                <Percent className="w-4 h-4 text-brand-500 dark:text-brand-400" />
-                {t("fire.inflation")}
-              </label>
-              <div className="relative">
-                <NumberInput
-                  value={inflation}
-                  onChange={(num) => {
-                    setInflation(Number.isNaN(num) ? 0 : num);
-                    userModified.current.inflation = true;
-                  }}
-                  className="form-input-lg font-semibold pr-8"
-                  placeholder="2"
-                  maximumFractionDigits={2}
-                  minimumFractionDigits={0}
-                  useGrouping={false}
-                />
-                <span className="absolute right-4 top-1/2 -translate-y-1/2 text-slate-400 font-medium">
-                  %
-                </span>
-              </div>
-            </div>
-
-            <div>
-              <label className="form-label !mb-2 flex items-center gap-2">
-                <Percent className="w-4 h-4 text-brand-500 dark:text-brand-400" />
-                {t("fire.withdrawal_rate")}
-              </label>
-              <div className="relative">
-                <NumberInput
-                  value={withdrawalRate}
-                  onChange={(num) => {
-                    setWithdrawalRate(Number.isNaN(num) ? 0 : num);
-                    userModified.current.withdrawalRate = true;
-                  }}
-                  className="form-input-lg font-semibold pr-8"
-                  placeholder="4"
-                  maximumFractionDigits={2}
-                  minimumFractionDigits={0}
-                  useGrouping={false}
-                />
-                <span className="absolute right-4 top-1/2 -translate-y-1/2 text-slate-400 font-medium">
-                  %
-                </span>
-              </div>
-            </div>
-
-            {/* Age & Timeline Section */}
-            <div className="pt-4 border-t border-slate-200 dark:border-slate-700">
-              <h3 className="text-sm font-bold text-slate-600 dark:text-slate-400 mb-4 flex items-center gap-2">
-                <User className="w-4 h-4" />
-                {t("fire.age_timeline")}
-              </h3>
-
-              <div className="space-y-4">
-                <div>
-                  <label className="form-label !mb-2 flex items-center gap-2">
-                    <User className="w-4 h-4 text-brand-500 dark:text-brand-400" />
-                    {t("fire.current_age")}
-                  </label>
-                  <div className="relative">
-                    <NumberInput
-                      value={currentAge}
-                      onChange={(num) => {
-                        setCurrentAge(Number.isNaN(num) ? 0 : Math.round(num));
-                        userModified.current.currentAge = true;
-                      }}
-                      className="form-input-lg font-semibold"
-                      placeholder="30"
-                      maximumFractionDigits={0}
-                      minimumFractionDigits={0}
-                      useGrouping={false}
-                    />
-                  </div>
-                </div>
-
-                <div>
-                  <label className="form-label !mb-2 flex items-center gap-2">
-                    <Target className="w-4 h-4 text-brand-500 dark:text-brand-400" />
-                    {t("fire.target_retirement_age")}
-                  </label>
-                  <div className="relative">
-                    <NumberInput
-                      value={retirementAge}
-                      onChange={(num) => {
-                        setRetirementAge(
-                          Number.isNaN(num) ? 0 : Math.round(num),
-                        );
-                        userModified.current.retirementAge = true;
-                      }}
-                      className="form-input-lg font-semibold"
-                      placeholder="65"
-                      maximumFractionDigits={0}
-                      minimumFractionDigits={0}
-                      useGrouping={false}
-                    />
-                  </div>
-                </div>
-
-                <div>
-                  <label className="form-label !mb-2 flex items-center gap-2">
-                    <Clock className="w-4 h-4 text-brand-500 dark:text-brand-400" />
-                    {t("fire.retirement_duration")}
-                  </label>
-                  <div className="relative">
-                    <NumberInput
-                      value={retirementDuration}
-                      onChange={(num) => {
-                        setRetirementDuration(
-                          Number.isNaN(num) ? 0 : Math.round(num),
-                        );
-                        userModified.current.retirementDuration = true;
-                      }}
-                      className="form-input-lg font-semibold"
-                      placeholder="30"
-                      maximumFractionDigits={0}
-                      minimumFractionDigits={0}
-                      useGrouping={false}
-                    />
-                    <span className="absolute right-4 top-1/2 -translate-y-1/2 text-slate-400 font-medium text-sm">
-                      {t("fire.years")}
-                    </span>
-                  </div>
-                </div>
-              </div>
-            </div>
-
-            {/* Advanced Options Toggle */}
-            <div className="pt-2">
-              <button
-                type="button"
-                onClick={() => setShowAdvanced(!showAdvanced)}
-                className="flex items-center gap-2 text-sm font-medium text-slate-600 dark:text-slate-400 hover:text-brand-600 dark:hover:text-brand-400 transition-colors"
-              >
-                {showAdvanced ? (
-                  <ChevronDown className="w-4 h-4" />
-                ) : (
-                  <ChevronRight className="w-4 h-4" />
-                )}
-                <Settings2 className="w-4 h-4" />
-                {showAdvanced
-                  ? t("fire.hide_advanced")
-                  : t("fire.show_advanced")}
-              </button>
-            </div>
-
-            {/* Advanced Monte Carlo Parameters */}
-            {showAdvanced && (
-              <div className="space-y-4 pt-2 pl-4 border-l-2 border-slate-200 dark:border-slate-700">
-                <p className="text-xs text-slate-500 dark:text-slate-400">
-                  {t("fire.advanced_description")}
-                </p>
-
-                <div>
-                  <label className="form-label !mb-2 flex items-center gap-2">
-                    <Activity className="w-4 h-4 text-brand-500 dark:text-brand-400" />
-                    {t("fire.return_volatility")}
-                  </label>
-                  <div className="relative">
-                    <NumberInput
-                      value={volatility}
-                      onChange={(num) => {
-                        setVolatility(Number.isNaN(num) ? 0 : num);
-                        userModified.current.volatility = true;
-                      }}
-                      className="form-input-lg font-semibold pr-8"
-                      placeholder="15"
-                      maximumFractionDigits={1}
-                      minimumFractionDigits={0}
-                      useGrouping={false}
-                    />
-                    <span className="absolute right-4 top-1/2 -translate-y-1/2 text-slate-400 font-medium">
-                      %
-                    </span>
-                  </div>
-                  <p className="text-xs text-slate-500 dark:text-slate-400 mt-1">
-                    {t("fire.volatility_hint")}
-                  </p>
-                </div>
-
-                <div>
-                  <label className="form-label !mb-2 flex items-center gap-2">
-                    <Activity className="w-4 h-4 text-brand-500 dark:text-brand-400" />
-                    {t("fire.simulation_count")}
-                  </label>
-                  <div className="relative">
-                    <NumberInput
-                      value={simulationCount}
-                      onChange={(num) => {
-                        setSimulationCount(
-                          Number.isNaN(num)
-                            ? 1000
-                            : Math.max(100, Math.min(10000, Math.round(num))),
-                        );
-                        userModified.current.simulationCount = true;
-                      }}
-                      className="form-input-lg font-semibold"
-                      placeholder="1000"
-                      maximumFractionDigits={0}
-                      minimumFractionDigits={0}
-                      useGrouping={false}
-                    />
-                  </div>
-                  <p className="text-xs text-slate-500 dark:text-slate-400 mt-1">
-                    {t("fire.simulation_count_hint")}
-                  </p>
-                </div>
-              </div>
-            )}
-          </div>
-        </div>
+        <FireInputsPanel
+          currentNetWorth={currentNetWorth}
+          setCurrentNetWorth={setCurrentNetWorth}
+          annualExpenses={annualExpenses}
+          setAnnualExpenses={setAnnualExpenses}
+          annualSavings={annualSavings}
+          setAnnualSavings={setAnnualSavings}
+          expectedReturn={expectedReturn}
+          setExpectedReturn={setExpectedReturn}
+          inflation={inflation}
+          setInflation={setInflation}
+          withdrawalRate={withdrawalRate}
+          setWithdrawalRate={setWithdrawalRate}
+          currentAge={currentAge}
+          setCurrentAge={setCurrentAge}
+          retirementAge={retirementAge}
+          setRetirementAge={setRetirementAge}
+          retirementDuration={retirementDuration}
+          setRetirementDuration={setRetirementDuration}
+          showAdvanced={showAdvanced}
+          setShowAdvanced={setShowAdvanced}
+          volatility={volatility}
+          setVolatility={setVolatility}
+          simulationCount={simulationCount}
+          setSimulationCount={setSimulationCount}
+          markUserModified={(field: string) => {
+            (userModified.current as Record<string, boolean>)[field] = true;
+          }}
+          resetToHistoric={resetToHistoric}
+        />
 
         {/* Results & Chart */}
-        <div className="lg:col-span-3 space-y-6">
-          {/* Key Metrics */}
-          <div className="grid grid-cols-1 md:grid-cols-2 xl:grid-cols-4 gap-4">
-            <div className="bg-gradient-to-br from-brand-50 to-brand-100 dark:from-brand-900/20 dark:to-brand-900/10 p-5 rounded-2xl shadow-md border-2 border-brand-200 dark:border-brand-800 flex items-center justify-between transition-all duration-300">
-              <div>
-                <p className="text-xs font-bold text-brand-700 dark:text-brand-300 uppercase tracking-wider mb-1">
-                  {t("fire.fire_number")}
-                </p>
-                <p className="text-2xl font-bold text-brand-900 dark:text-brand-100">
-                  <MaskedNumber
-                    value={fireNumber}
-                    options={{
-                      style: "currency",
-                      maximumFractionDigits: 0,
-                      minimumFractionDigits: 0,
-                    }}
-                  />
-                </p>
-              </div>
-              <div className="bg-brand-500 dark:bg-brand-600 p-3 rounded-xl shadow-lg">
-                <TrendingUp className="w-6 h-6 text-white" />
-              </div>
-            </div>
-
-            <div className="bg-gradient-to-br from-brand-50 to-brand-100 dark:from-brand-900/20 dark:to-brand-900/10 p-5 rounded-2xl shadow-md border-2 border-brand-200 dark:border-brand-800 flex items-center justify-between transition-all duration-300">
-              <div>
-                <p className="text-xs font-bold text-brand-700 dark:text-brand-300 uppercase tracking-wider mb-1">
-                  {t("fire.time_to_fire")}
-                </p>
-                {neverReached ? (
-                  <p className="text-lg font-medium text-brand-900 dark:text-brand-100">
-                    {t("fire.never_retire")}
-                  </p>
-                ) : (
-                  <p className="text-2xl font-bold text-brand-900 dark:text-brand-100">
-                    {yearsToFire} {t("fire.years")}
-                  </p>
-                )}
-              </div>
-              <div className="bg-brand-500 dark:bg-brand-600 p-3 rounded-xl shadow-lg">
-                <Calendar className="w-6 h-6 text-white" />
-              </div>
-            </div>
-
-            <div className="bg-gradient-to-br from-brand-50 to-brand-100 dark:from-brand-900/20 dark:to-brand-900/10 p-5 rounded-2xl shadow-md border-2 border-brand-200 dark:border-brand-800 flex items-center justify-between transition-all duration-300">
-              <div>
-                <p className="text-xs font-bold text-brand-700 dark:text-brand-300 uppercase tracking-wider mb-1">
-                  {t("fire.retirement_age")}
-                </p>
-                {neverReached || !fireAge ? (
-                  <p className="text-lg font-medium text-brand-900 dark:text-brand-100">
-                    —
-                  </p>
-                ) : (
-                  <p className="text-2xl font-bold text-brand-900 dark:text-brand-100">
-                    {t("fire.age_value", { age: fireAge })}
-                  </p>
-                )}
-              </div>
-              <div className="bg-brand-500 dark:bg-brand-600 p-3 rounded-xl shadow-lg">
-                <User className="w-6 h-6 text-white" />
-              </div>
-            </div>
-
-            <div className="bg-gradient-to-br from-brand-50 to-brand-100 dark:from-brand-900/20 dark:to-brand-900/10 p-5 rounded-2xl shadow-md border-2 border-brand-200 dark:border-brand-800 flex items-center justify-between transition-all duration-300">
-              <div>
-                <p className="text-xs font-bold text-brand-700 dark:text-brand-300 uppercase tracking-wider mb-1">
-                  {t("fire.success_rate")}
-                </p>
-                <p className="text-2xl font-bold text-brand-900 dark:text-brand-100">
-                  {monteCarloResult
-                    ? `${monteCarloResult.successRate.toFixed(1)}%`
-                    : "—"}
-                </p>
-                <p className="text-xs text-slate-500 dark:text-slate-400 mt-1">
-                  {t("fire.monte_carlo")}
-                </p>
-              </div>
-              <div className="bg-brand-500 dark:bg-brand-600 p-3 rounded-xl shadow-lg">
-                <Activity className="w-6 h-6 text-white" />
-              </div>
-            </div>
-          </div>
-
-          {/* Chart */}
-          <div className="bg-white dark:bg-slate-800 p-6 rounded-2xl shadow-md border border-slate-200 dark:border-slate-700 flex-1 min-h-[400px] hover:shadow-lg transition-shadow duration-300">
-            <div className="flex items-center justify-between mb-2">
-              <div>
-                <h3 className="text-lg font-semibold text-slate-700 dark:text-slate-200">
-                  {t("fire.projection")}
-                </h3>
-                <p className="text-sm text-slate-500 dark:text-slate-400">
-                  {t("fire.projection_subtitle")}
-                </p>
-              </div>
-              {monteCarloResult && (
-                <div className="text-right">
-                  <p className="text-xs text-slate-500 dark:text-slate-400">
-                    {t("fire.simulations_run", {
-                      count: monteCarloResult.simulationCount,
-                    })}
-                  </p>
-                </div>
-              )}
-            </div>
-            <div className="h-[350px]">
-              <Line
-                data={chartData}
-                options={{
-                  responsive: true,
-                  maintainAspectRatio: false,
-                  interaction: {
-                    mode: "index",
-                    intersect: false,
-                  },
-                  plugins: {
-                    legend: {
-                      position: "top",
-                      labels: {
-                        color: chartColors.text,
-                        filter: (item) => {
-                          // Hide some labels to reduce clutter
-                          const hiddenLabels = [
-                            t("fire.percentile_10"),
-                            t("fire.percentile_25"),
-                            t("fire.percentile_75"),
-                          ];
-                          return !hiddenLabels.includes(item.text);
-                        },
-                      },
-                    },
-                    tooltip: {
-                      backgroundColor: chartColors.tooltipBg,
-                      titleColor: chartColors.tooltipText,
-                      bodyColor: chartColors.tooltipText,
-                      padding: 12,
-                      cornerRadius: 8,
-                      callbacks: {
-                        labelColor: function (context) {
-                          const dataset = context.dataset;
-                          const tooltipBg = chartColors.tooltipBg;
-
-                          // Always use the tooltip background as the fill color for the label box
-                          // This ensures a "hollow" look matching the line style, avoiding issues
-                          // with semi-transparent fills (0.1 opacity) looking washed out or "white".
-                          return {
-                            borderColor: dataset.borderColor as string,
-                            backgroundColor: tooltipBg,
-                            borderWidth: 2,
-                          };
-                        },
-                        label: function (context) {
-                          let label = context.dataset.label || "";
-                          if (label) {
-                            label += ": ";
-                          }
-
-                          const value =
-                            (context.parsed &&
-                              (context.parsed.y ?? context.parsed)) ??
-                            context.raw ??
-                            (context.dataset &&
-                            context.dataset.data &&
-                            context.dataIndex != null
-                              ? context.dataset.data[context.dataIndex]
-                              : undefined);
-
-                          if (
-                            value !== undefined &&
-                            value !== null &&
-                            !Number.isNaN(Number(value))
-                          ) {
-                            label += formatNumber(Number(value), {
-                              style: "currency",
-                              ignorePrivacy: true,
-                            });
-                          }
-
-                          return label;
-                        },
-                      },
-                    },
-                  },
-                  scales: {
-                    y: {
-                      beginAtZero: true,
-                      grid: {
-                        color: chartColors.grid,
-                      },
-                      ticks: {
-                        color: chartColors.text,
-                        callback: function (value) {
-                          const num = Number(value);
-                          if (Number.isNaN(num)) return value;
-                          return formatNumber(num, {
-                            style: "currency",
-                          });
-                        },
-                      },
-                    },
-                    x: {
-                      grid: {
-                        color: chartColors.grid,
-                      },
-                      ticks: {
-                        color: chartColors.text,
-                        maxTicksLimit: 10,
-                      },
-                    },
-                  },
-                }}
-              />
-            </div>
-
-            {/* Chart Legend Explanation */}
-            <div className="mt-4 pt-4 border-t border-slate-200 dark:border-slate-700">
-              <p className="text-xs text-slate-500 dark:text-slate-400">
-                {t("fire.chart_legend_explanation")}
-              </p>
-            </div>
-          </div>
-        </div>
+        <FireResultsPanel
+          fireNumber={fireNumber}
+          yearsToFire={yearsToFire}
+          neverReached={neverReached}
+          fireAge={fireAge}
+          monteCarloResult={monteCarloResult}
+          chartData={chartData}
+        />
       </div>
     </div>
   );
